\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\description{Convert Celsius to Fahrenheit}
\usage{
C_to_F(value)
}
\arguments{
  \item{value}{A numeric value representing degrees Celsius to be converted.}
}
\value{
A numeric value representing the equivalent in Fahrenheit.
}
\examples{
C_to_F(0)
}
