% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{Numeric value in Celsius.}
}
\value{
Numeric value in Fahrenheit.
}
\description{
This function converts a given temperature in Celsius to Fahrenheit.
}
\examples{
celsius_to_fahrenheit(0)
}
