% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{centimeters_to_inches}
\alias{centimeters_to_inches}
\title{Convert Centimeters to Inches}
\usage{
centimeters_to_inches(centimeters)
}
\arguments{
\item{centimeters}{Numeric value in centimeters.}
}
\value{
Numeric value in inches.
}
\description{
This function converts a given length in centimeters to inches.
}
\examples{
centimeters_to_inches(2.54)
}
