% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_atm}
\alias{cisnienie_atm}
\title{Konwersja cisnienia (cisnienie atmosferyczne)}
\usage{
cisnienie_atm(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc cisnienia atmosferycznego}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji cisnienia atmosferycznego na pozostale jednostki cisnienia
}
\examples{
cisnienie_atm(20)
cisnienie_atm(c(40, 95))
}

