% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_bar}
\alias{cisnienie_bar}
\title{Konwersja cisnienia (bar)}
\usage{
cisnienie_bar(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc barow}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji barow na pozostale jednostki cisnienia
}
\examples{
cisnienie_bar(20)
cisnienie_bar(c(40, 95))
}
