% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_inHg}
\alias{cisnienie_inHg}
\title{Konwersja cisnienia (cal slupa rteci)}
\usage{
cisnienie_inHg(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc cali slupa rteci}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji cali slupa rteci na pozostale jednostki cisnienia
}
\examples{
cisnienie_inHg(20)
cisnienie_inHg(c(40, 95))
}
