% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_inWg}
\alias{cisnienie_inWg}
\title{Konwersja cisnienia (cal slupa wody)}
\usage{
cisnienie_inWg(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc cali slupa wody}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji cali slupa wody na pozostale jednostki cisnienia
}
\examples{
cisnienie_inWg(20)
cisnienie_inWg(c(40, 95))
}
