% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_kPa}
\alias{cisnienie_kPa}
\title{Konwersja cisnienia (kilopaskale)}
\usage{
cisnienie_kPa(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilopaskali}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilopaskali na pozostale jednostki cisnienia
}
\examples{
cisnienie_kPa(20)
cisnienie_kPa(c(40, 95))
}
