% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_psi}
\alias{cisnienie_psi}
\title{Konwersja cisnienia (funty na cal kwadratowy)}
\usage{
cisnienie_psi(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc funtow na cal kwadratowy}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji funtow na cal kwadratowy na pozostale jednostki cisnienia
}
\examples{
cisnienie_psi(20)
cisnienie_psi(c(40, 95))
}
