\name{cm_to_inches}
\alias{cm_to_inches}
\title{Centimetres to Inches}
\description{Convert centimetres to inches}
\usage{
cm_to_inches(value)
}
\arguments{
  \item{value}{A numeric value representing centimetres to be converted.}
}
\value{
A numeric value representing the equivalent in inches.
}
\examples{
cm_to_inches(1)
}
