% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_volume}
\alias{convert_volume}
\title{Convert volume}
\usage{
convert_volume(value, from, to)
}
\arguments{
\item{value}{Numeric volume value}

\item{from}{Unit: "ml", "l", "tsp", "tbsp", "cup"}

\item{to}{Unit: "ml", "l", "tsp", "tbsp", "cup"}
}
\value{
Converted volume
}
\description{
Convert volume
}
