% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_weight}
\alias{convert_weight}
\title{Convert weight}
\usage{
convert_weight(value, from, to)
}
\arguments{
\item{value}{Numeric weight value}

\item{from}{Unit: "g", "kg", "lb", "oz"}

\item{to}{Unit: "g", "kg", "lb", "oz"}
}
\value{
Converted weight
}
\description{
Convert weight
}
