% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_converter.R
\name{dist_converter}
\alias{dist_converter}
\title{Convert between distances}
\usage{
dist_converter(dist_value, unit_from, unit_to)
}
\arguments{
\item{dist_value}{A numerical distance value to be converted.}

\item{unit_from}{A character string of the distance unit to convert from}

\item{unit_to}{A character string of the distance unit to convert to}
}
\value{
A numeric distance value in the unit specified as \code{unit_to}.
}
\description{
Performs the conversion based on the specified \code{unit_from} and \code{unit_to} values.
}
