% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_data.R
\docType{data}
\name{distance_data}
\alias{distance_data}
\title{Distance values and units}
\format{
Data frame with two columns and 1000 rows
\describe{
\item{value}{Numeric distance value}
\item{unit}{Distance unit}
}
}
\usage{
distance_data
}
\description{
Distance values and corresponding unit (feet, meters, or inches)
}
\keyword{datasets}
