% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_cm}
\alias{dlugosc_cm}
\title{Konwersja dlugosci (centymetry)}
\usage{
dlugosc_cm(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc centymetrow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji centymetrow na pozostale jednostki dlugosci
}
\examples{
dlugosc_cm(20)
dlugosc_cm(c(40, 95))
}
