% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_in}
\alias{dlugosc_in}
\title{Konwersja dlugosci (cale)}
\usage{
dlugosc_in(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc cali}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji cali na pozostale jednostki dlugosci
}
\examples{
dlugosc_in(20)
dlugosc_in(c(40, 95))
}
