% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_km}
\alias{dlugosc_km}
\title{Konwersja dlugosci (kilometry)}
\usage{
dlugosc_km(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilometrow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilometrow na pozostale jednostki dlugosci
}
\examples{
dlugosc_km(20)
dlugosc_km(c(40, 95))
}
