% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_ly}
\alias{dlugosc_ly}
\title{Konwersja dlugosci (lata swietlne)}
\usage{
dlugosc_ly(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc lat swietlnych}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji lat swietlnych na pozostale jednostki dlugosci
}
\examples{
dlugosc_ly(20)
dlugosc_ly(c(40, 95))
}
