% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_m}
\alias{dlugosc_m}
\title{Konwersja dlugosci (metry)}
\usage{
dlugosc_m(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc metrow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji metrow na pozostale jednostki dlugosci
}
\examples{
dlugosc_m(20)
dlugosc_m(c(40, 95))
}
