% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_mila}
\alias{dlugosc_mila}
\title{Konwersja dlugosci (mila)}
\usage{
dlugosc_mila(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc mil}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji mil na pozostale jednostki dlugosci
}
\examples{
dlugosc_mila(20)
dlugosc_mila(c(40, 95))
}
