\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius}
\description{Convert Fahrenheit to Celsius}
\usage{
F_to_C(value)
}
\arguments{
  \item{value}{A numeric value representing degrees Fahrenheit to be converted.}
}
\value{
A numeric value representing the equivalent in Celsius.
}
\examples{
F_to_C(0)
}
