% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{Numeric value in Fahrenheit.}
}
\value{
Numeric value in Celsius.
}
\description{
This function converts a given temperature in Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32)
}
