% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{grams_to_kilograms}
\alias{grams_to_kilograms}
\title{Convert Grams to Kilograms}
\usage{
grams_to_kilograms(grams)
}
\arguments{
\item{grams}{Numeric value in grams.}
}
\value{
Numeric value in kilograms.
}
\description{
This function converts a given weight in grams to kilograms.
}
\examples{
grams_to_kilograms(1000)
}
