% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{inches_to_centimeters}
\alias{inches_to_centimeters}
\title{Convert Inches to Centimeters}
\usage{
inches_to_centimeters(inches)
}
\arguments{
\item{inches}{Numeric value in inches.}
}
\value{
Numeric value in centimeters.
}
\description{
This function converts a given length in inches to centimeters.
}
\examples{
inches_to_centimeters(1)
}
