\name{inches_to_cm}
\alias{inches_to_cm}
\title{Inches to Centimetres}
\description{Convert inches to centimetres}
\usage{
inches_to_cm(value)
}
\arguments{
  \item{value}{A numeric value representing inches to be converted.}
}
\value{
A numeric value representing the equivalent in centimetres.
}
\examples{
inches_to_cm(1)
}
