% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{kilograms_to_grams}
\alias{kilograms_to_grams}
\title{Convert Kilograms to Grams}
\usage{
kilograms_to_grams(kilograms)
}
\arguments{
\item{kilograms}{Numeric value in kilograms.}
}
\value{
Numeric value in grams.
}
\description{
This function converts a given weight in kilograms to grams.
}
\examples{
kilograms_to_grams(1)
}
