% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{kilometers_to_miles}
\alias{kilometers_to_miles}
\title{Convert Kilometers to Miles}
\usage{
kilometers_to_miles(kilometers)
}
\arguments{
\item{kilometers}{Numeric value in kilometers.}
}
\value{
Numeric value in miles.
}
\description{
This function converts a given value in kilometers to miles.
}
\examples{
kilometers_to_miles(1.60934)
}
