% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{liters_to_milliliters}
\alias{liters_to_milliliters}
\title{Convert Liters to Milliliters}
\usage{
liters_to_milliliters(liters)
}
\arguments{
\item{liters}{Numeric value in liters.}
}
\value{
Numeric value in milliliters.
}
\description{
This function converts a given volume in liters to milliliters.
}
\examples{
liters_to_milliliters(1)
}
