% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_ct}
\alias{masa_ct}
\title{Konwersja masy (karaty)}
\usage{
masa_ct(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc karatow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji karatow na pozostale jednostki masy
}
\examples{
masa_ct(20)
masa_ct(c(40, 95))
}
