% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_gr}
\alias{masa_gr}
\title{Konwersja masy (grany)}
\usage{
masa_gr(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc granow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji granow na pozostale jednostki masy
}
\examples{
masa_gr(20)
masa_gr(c(40, 95))
}
