% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_kg}
\alias{masa_kg}
\title{Konwersja masy (kilogramy)}
\usage{
masa_kg(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilogramow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilogramow na pozostale jednostki masy
}
\examples{
masa_kg(20)
masa_kg(c(40, 95))
}
