% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_t}
\alias{masa_t}
\title{Konwersja masy (tony)}
\usage{
masa_t(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc ton}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji ton na pozostale jednostki masy
}
\examples{
masa_t(20)
masa_t(c(40, 95))
}
