% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_ton_uk}
\alias{masa_ton_uk}
\title{Konwersja masy (tony brytyjskie)}
\usage{
masa_ton_uk(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc ton brytyjskich}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji ton brytyjskich na pozostale jednostki masy
}
\examples{
masa_ton_uk(20)
masa_ton_uk(c(40, 95))
}
