% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_u}
\alias{masa_u}
\title{Konwersja masy (masa atomowa)}
\usage{
masa_u(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc masy atomowej}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji masy atomowej na pozostale jednostki masy
}
\examples{
masa_u(20)
masa_u(c(40, 95))
}
