% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{miles_to_kilometers}
\alias{miles_to_kilometers}
\title{Convert Miles to Kilometers}
\usage{
miles_to_kilometers(miles)
}
\arguments{
\item{miles}{Numeric value in miles.}
}
\value{
Numeric value in kilometers.
}
\description{
This function converts a given value in miles to kilometers.
}
\examples{
miles_to_kilometers(1)
}
