% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_in3}
\alias{objetosc_in3}
\title{Konwersja objetosci (cale szescienne)}
\usage{
objetosc_in3(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc cali szesciennych}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji cali szesciennych na pozostale jednostki objetosci
}
\examples{
objetosc_in3(20)
objetosc_in3(c(40, 95))
}
