% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_L}
\alias{objetosc_L}
\title{Konwersja objetosci (litr)}
\usage{
objetosc_L(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc litrow}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji litrow na pozostale jednostki objetosci
}
\examples{
objetosc_L(20)
objetosc_L(c(40, 95))
}
