% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_ml}
\alias{objetosc_ml}
\title{Konwersja objetosci (mililitry)}
\usage{
objetosc_ml(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc mililitrow}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji mililitrow na pozostale jednostki objetosci
}
\examples{
objetosc_ml(20)
objetosc_ml(c(40, 95))
}
