% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_pinta_uk}
\alias{objetosc_pinta_uk}
\title{Konwersja objetosci (pinta brytyjska)}
\usage{
objetosc_pinta_uk(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc pinty brytyjskiej}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji pinty brytyjskiej na pozostale jednostki objetosci
}
\examples{
objetosc_pinta_uk(20)
objetosc_pinta_uk(c(40, 95))
}
