% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_pinta_us}
\alias{objetosc_pinta_us}
\title{Konwersja objetosci (pinta amerykanska)}
\usage{
objetosc_pinta_us(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc pinty amerykanskiej}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji pinty amerykanskiej na pozostale jednostki objetosci
}
\examples{
objetosc_pinta_us(20)
objetosc_pinta_us(c(40, 95))
}
