% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_uncja_uk}
\alias{objetosc_uncja_uk}
\title{Konwersja objetosci (uncja brytyjska)}
\usage{
objetosc_uncja_uk(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc uncji brytyjskich}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji uncji brytyjskich na pozostale jednostki objetosci
}
\examples{
objetosc_uncja_uk(20)
objetosc_uncja_uk(c(40, 95))
}
