% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_cm2}
\alias{powierzchnia_cm2}
\title{Konwersja powierzchni (centymetry kwadratowe)}
\usage{
powierzchnia_cm2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc centymetrow kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji centymetrow kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_cm2(20)
powierzchnia_cm2(c(40, 95))
}
