% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_ft2}
\alias{powierzchnia_ft2}
\title{Konwersja powierzchni (stopy kwadratowe)}
\usage{
powierzchnia_ft2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stop kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stop kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_ft2(20)
powierzchnia_ft2(c(40, 95))
}
