% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_ha}
\alias{powierzchnia_ha}
\title{Konwersja powierzchni (hektary)}
\usage{
powierzchnia_ha(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc hektarow}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji hektarow na pozostale jednostki powierzchni
}
\examples{
powierzchnia_ha(20)
powierzchnia_ha(c(40, 95))
}
