% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_km2}
\alias{powierzchnia_km2}
\title{Konwersja powierzchni (kilometry kwadratowe)}
\usage{
powierzchnia_km2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilometrow kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilometrow kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_km2(20)
powierzchnia_km2(c(40, 95))
}
