% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_m2}
\alias{powierzchnia_m2}
\title{Konwersja powierzchni (metry kwadratowe)}
\usage{
powierzchnia_m2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc metrow kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji metrow kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_m2(20)
powierzchnia_m2(c(40, 95))
}
