% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_yd2}
\alias{powierzchnia_yd2}
\title{Konwersja powierzchni (jardy kwadratowe)}
\usage{
powierzchnia_yd2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc jardow kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji jardow kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_yd2(20)
powierzchnia_yd2(c(40, 95))
}
