% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_fts}
\alias{predkosc_fts}
\title{Konwersja predkosci (stopy na sekunde)}
\usage{
predkosc_fts(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stop na sekunde}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stop na sekunde na pozostale jednostki predkosci
}
\examples{
predkosc_fts(20)
predkosc_fts(c(40, 95))
}
