% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_kmmin}
\alias{predkosc_kmmin}
\title{Konwersja predkosci (kilometry na minute)}
\usage{
predkosc_kmmin(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilometrow na minute}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilometrow na minute na pozostale jednostki predkosci
}
\examples{
predkosc_kmmin(20)
predkosc_kmmin(c(40, 95))
}
