% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_kms}
\alias{predkosc_kms}
\title{Konwersja predkosci (kkilometry na sekunde)}
\usage{
predkosc_kms(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilometrow na sekunde}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilometrow na sekunde na pozostale jednostki predkosci
}
\examples{
predkosc_kms(20)
predkosc_kms(c(40, 95))
}
