% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_kn}
\alias{predkosc_kn}
\title{Konwersja predkosci (wezly)}
\usage{
predkosc_kn(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc wezlow}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji wezlow na pozostale jednostki predkosci
}
\examples{
predkosc_kn(20)
predkosc_kn(c(40, 95))
}
