% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_Mach}
\alias{predkosc_Mach}
\title{Konwersja predkosci (machy)}
\usage{
predkosc_Mach(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc Mach}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji Mach na pozostale jednostki predkosci
}
\examples{
predkosc_Mach(20)
predkosc_Mach(c(40, 95))
}
