% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_mmin}
\alias{predkosc_mmin}
\title{Konwersja predkosci (metry na minute)}
\usage{
predkosc_mmin(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc metrow na minute}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji metrow na minute na pozostale jednostki predkosci
}
\examples{
predkosc_mmin(20)
predkosc_mmin(c(40, 95))
}
