% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_converter.R
\name{temp_converter}
\alias{temp_converter}
\title{Convert between Fahrenheit and Celsius temperatures}
\usage{
temp_converter(value, unit_from = "Celsius", unit_to = "Fahrenheit")
}
\arguments{
\item{value}{A numerical value of the temperature to be converted.}

\item{unit_from}{A character string of the temperature unit to convert from
(default is "Celsius").}

\item{unit_to}{A character string of the temperature unit to convert to
(default is "Fahrenheit").}
}
\value{
A numeric temperature value in the unit specified as \code{unit_to}.
}
\description{
\code{temp_converter()} performs the conversion based on the specified \code{unit_from}
and \code{unit_to} values. If the conversion is not possible due to invalid units or
if \code{unit_from} and \code{unit_to} are the same, appropriate warnings or errors
are raised.
}
