% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{temperatura_F}
\alias{temperatura_F}
\title{Konwersja temperatury (stopnie Fahrenheita)}
\usage{
temperatura_F(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stopni Celcjusza}
}
\value{
data frame z pozostalymi jednostkami temperatury i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stopni Fahrenheita na pozostale jednostki temperatury
}
\examples{
temperatura_F(20)
temperatura_F(c(40, 95))
}
