% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converter.R
\name{time_converter}
\alias{time_converter}
\title{Convert between hours, minutes, and seconds}
\usage{
time_converter(time, unit_from, unit_to)
}
\arguments{
\item{time}{A numerical value of the time unit to be converted.}

\item{unit_from}{A character string of the time unit to convert from}

\item{unit_to}{A character string of the time unit to convert to}
}
\value{
A numeric time value in the unit specified as \code{unit_to}.
}
\description{
\code{time_converter()} performs the conversion based on the specified \code{unit_from}
and \code{unit_to} values.
}
