% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_converter.R
\name{weight_converter}
\alias{weight_converter}
\title{Performs the weight conversion based on the specified \code{unit_from} and \code{unit_to} values.}
\usage{
weight_converter(weight_value, unit_from, unit_to)
}
\arguments{
\item{weight_value}{A numerical weight value to be converted.}

\item{unit_from}{A character string of the distance unit to convert from. Options are \code{"gram"},
\code{"kilogram"}, \code{"pound"}, and \code{"ounce"}.}

\item{unit_to}{A character string of the distance unit to convert to. Options are \code{"gram"},
\code{"kilogram"}, \code{"pound"}, and \code{"ounce"}.}
}
\value{
A numeric weight value in the unit specified as \code{unit_to}.
}
\description{
Performs the weight conversion based on the specified \code{unit_from} and \code{unit_to} values.
}
