% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_converter.R
\name{dist_converter}
\alias{dist_converter}
\title{Convert between distances}
\usage{
dist_converter(dist_value, unit_from, unit_to)
}
\arguments{
\item{dist_value}{value to be converted}

\item{unit_from}{which unit you're converting from}

\item{unit_to}{which unit you're converting to}

\item{dist_converter()}{A numerical distance value to be converted.}
}
\value{
A numeric distance value in the unit specified as unit_to.
}
\description{
Performs the conversion based on specified unit_from and unit_to values.
}
\examples{
# Meters to feet
dist_converter(4, "meters", "feet")
# Feet to meters
dist_converter(10.5, "feet", "meters")
\dontrun{
# Feet to feet (gives warning)
dist_converter(100, "feet", "feet")
# Non-numeric `dist_value` gives error
dist_converter("short", "feet", "meters")
}
}
