% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_converter.R
\name{dist_converter}
\alias{dist_converter}
\title{Convert Distance Between Feet and Meters}
\usage{
dist_converter(dist_value, unit_from, unit_to)
}
\arguments{
\item{dist_value}{A numeric value representing the distance to be converted.}

\item{unit_from}{A character string specifying the unit of the input distance value. Must be either "feet" or "meters".}

\item{unit_to}{A character string specifying the unit of the output distance value. Must be either "feet" or "meters".}
}
\value{
The converted distance value. If `unit_from` and `unit_to` are the same, the original `dist_value` is returned with a warning.
}
\description{
This function converts a distance value between feet and meters.
}
\examples{
dist_converter(10, "feet", "meters") # Converts 10 feet to meters
dist_converter(3.048, "meters", "feet") # Converts 3.048 meters to feet
dist_converter(5, "feet", "feet") # Returns 5 with a warning

}
