% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_converter.R
\name{temp_converter}
\alias{temp_converter}
\title{Convert Temperature Between Celsius and Fahrenheit}
\usage{
temp_converter(value, unit_from = "Celsius", unit_to = "Fahrenheit")
}
\arguments{
\item{value}{A numeric value representing the temperature to be converted.}

\item{unit_from}{A character string specifying the unit of the input temperature value. Must be either "Celsius" or "Fahrenheit". Default is "Celsius".}

\item{unit_to}{A character string specifying the unit of the output temperature value. Must be either "Celsius" or "Fahrenheit". Default is "Fahrenheit".}
}
\value{
The converted temperature value. If `unit_from` and `unit_to` are the same, the original `value` is returned with a warning.
}
\description{
This function converts a temperature value between Celsius and Fahrenheit.
}
\examples{
temp_converter(0, "Celsius", "Fahrenheit") # Converts 0 Celsius to Fahrenheit
temp_converter(32, "Fahrenheit", "Celsius") # Converts 32 Fahrenheit to Celsius
temp_converter(100, "Celsius", "Celsius") # Returns 100 with a warning

}
\seealso{
\link[base]{stop}, \link[base]{warning}
}
