% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converter.R
\name{time_converter}
\alias{time_converter}
\title{Convert Time Between Hours, Minutes, and Seconds}
\usage{
time_converter(time, unit_from, unit_to)
}
\arguments{
\item{time}{A numeric value representing the time to be converted.}

\item{unit_from}{A character string specifying the unit of the input time value. Must be one of "hours", "minutes", or "seconds".}

\item{unit_to}{A character string specifying the unit of the output time value. Must be one of "hours", "minutes", or "seconds".}
}
\value{
The converted time value. If `unit_from` and `unit_to` are the same, the original `time` is returned.
}
\description{
This function converts a time value between hours, minutes, and seconds.
}
\examples{
time_converter(2, "hours", "minutes") # Converts 2 hours to minutes
time_converter(120, "minutes", "hours") # Converts 120 minutes to hours
time_converter(3600, "seconds", "hours") # Converts 3600 seconds to hours
time_converter(45, "minutes", "minutes") # Returns 45 (no conversion needed)

}
