% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_converter.R
\name{weight_converter}
\alias{weight_converter}
\title{Convert Weight Between Grams, Kilograms, Pounds, and Ounces}
\usage{
weight_converter(weight_value, unit_from, unit_to)
}
\arguments{
\item{weight_value}{A numeric value representing the weight to be converted.}

\item{unit_from}{A character string specifying the unit of the input weight value. Must be one of "gram", "kilogram", "pound", or "ounce".}

\item{unit_to}{A character string specifying the unit of the output weight value. Must be one of "gram", "kilogram", "pound", or "ounce".}
}
\value{
The converted weight value.
}
\description{
This function converts a weight value between grams, kilograms, pounds, and ounces.
}
\examples{
weight_converter(1000, "gram", "kilogram") # Converts 1000 grams to kilograms
weight_converter(2.2, "pound", "kilogram") # Converts 2.2 pounds to kilograms
weight_converter(500, "ounce", "gram") # Converts 500 ounces to grams

}
